+++ src/app/views/pages/login/login.component.ts
@@ -1,49 +1,19 @@
 import { Component } from '@angular/core';
-import { NgStyle } from '@angular/common';
-import { IconDirective } from '@coreui/icons-angular';
-import {
-  ButtonDirective,
-  CardBodyComponent,
-  CardComponent,
-  CardGroupComponent,
-  ColComponent,
-  ContainerComponent,
-  FormControlDirective,
-  FormDirective,
-  InputGroupComponent,
-  InputGroupTextDirective,
-  RowComponent,
-} from '@coreui/angular';
-import {
-  FormControl,
-  FormGroup,
-  FormsModule,
-  ReactiveFormsModule,
-  Validators,
-} from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule, FormGroup, FormControl, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { AuthService } from '../../../core/services/auth.service';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { ToastrService } from 'ngx-toastr';
+import { AuthService } from '../../../core/services/auth.service';
 
 @Component({
   selector: 'app-login',
+  standalone: true,
   templateUrl: './login.component.html',
+  styleUrls: ['./login.component.scss'],
   imports: [
-    ContainerComponent,
-    FormsModule,
+    CommonModule,
     ReactiveFormsModule,
-    RowComponent,
-    ColComponent,
-    CardGroupComponent,
-    CardComponent,
-    CardBodyComponent,
-    FormDirective,
-    InputGroupComponent,
-    InputGroupTextDirective,
-    IconDirective,
-    FormControlDirective,
-    ButtonDirective,
     HttpClientModule,
   ],
   providers: [AuthService, HttpClient],
@@ -53,8 +23,10 @@ export class LoginComponent {
     email: new FormControl('', [Validators.required, Validators.email]),
     password: new FormControl('', [Validators.required]),
   });
-  loginLoading: boolean = false;
+
+  loginLoading = false;
   timeOutmessage = 5000;
+
   constructor(
     private router: Router,
     private authService: AuthService,
@@ -69,14 +41,15 @@ export class LoginComponent {
       this.formControl.markAllAsTouched();
       return;
     }
+
     this.authService
       .login(
         this.formControl.get('email')?.value || '',
         this.formControl.get('password')?.value || ''
       )
       .subscribe({
-        next: async (userResponse: any) => {
-          this.loginLoading = true;
+        next: (userResponse: any) => {
+          this.loginLoading = false;
           this.toastr.success('Sesión iniciada con éxito', 'Realizado', {
             timeOut: this.timeOutmessage,
             closeButton: true,